telegram-bot/requirements.txt
boto3
six
python-telegram-bot==13.14
ydb
requests